backend/pyproject.toml
[tool.poetry]
name = "memoria-backend"
version = "0.1.0"
description = "Memoria API"
authors = ["Memoria Team <dev@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.114.0"
uvicorn = {version = "^0.30.0", extras=["standard"]}
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.1"
redis = "^5.0.7"
httpx = "^0.27.0"
sentence-transformers = "^2.2.2"
nltk = "^3.8.1"
spacy = "^3.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


